+++ src/main/java/fr/raksrinana/fallingtree/ForgeEventSubscriber.java
@@ -2,6 +2,7 @@ package fr.raksrinana.fallingtree;
 
 import fr.raksrinana.fallingtree.config.Config;
 import fr.raksrinana.fallingtree.tree.TreeHandler;
+import io.netty.util.internal.ConcurrentSet;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.player.PlayerEntity;
@@ -10,15 +11,21 @@ import net.minecraft.util.Direction;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TranslationTextComponent;
-import net.minecraft.world.IWorld;
 import net.minecraft.world.server.ServerWorld;
+import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.fml.common.Mod;
 import javax.annotation.Nonnull;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.Set;
 
 @Mod.EventBusSubscriber(modid = FallingTree.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public final class ForgeEventSubscriber{
+	private static final Set<ScheduledLeafBreak> scheduledLeavesBreaking = new ConcurrentSet<>();
+	
 	@SubscribeEvent
 	public static void onBlockBreakEvent(@Nonnull BlockEvent.BreakEvent event){
 		if(!event.isCanceled() && !event.getWorld().isRemote()){
@@ -50,7 +57,7 @@ public final class ForgeEventSubscriber{
 	@SubscribeEvent
 	public static void onNeighborNotifyEvent(BlockEvent.NeighborNotifyEvent event){
 		if(Config.COMMON.breakLeaves.get() && !event.getWorld().isRemote()){
-			IWorld world = event.getWorld();
+			ServerWorld world = (ServerWorld) event.getWorld();
 			BlockState eventState = event.getState();
 			Block eventBlock = eventState.getBlock();
 			BlockPos eventPos = event.getPos();
@@ -60,9 +67,38 @@ public final class ForgeEventSubscriber{
 					if(world.isBlockLoaded(neighborPos)){
 						BlockState neighborState = event.getWorld().getBlockState(neighborPos);
 						if(BlockTags.LEAVES.contains(neighborState.getBlock())){
-							neighborState.getBlock().randomTick(neighborState, (ServerWorld) world, neighborPos, world.getRandom());
+							scheduledLeavesBreaking.add(new ScheduledLeafBreak(world, neighborPos, 4));
 						}
 					}
+				}
+			}
+		}
+	}
+	
+	@SubscribeEvent
+	public static void onServerTick(TickEvent.ServerTickEvent event){
+		if(event.side == LogicalSide.SERVER && event.phase == TickEvent.Phase.END){
+			Iterator<ScheduledLeafBreak> leavesBreak = scheduledLeavesBreaking.iterator();
+			while(leavesBreak.hasNext()){
+				ScheduledLeafBreak scheduledLeafBreak = leavesBreak.next();
+				ServerWorld world = scheduledLeafBreak.getWorld();
+				if(scheduledLeafBreak.getRemainingTicks() <= 0){
+					if(world.isBlockLoaded(scheduledLeafBreak.getBlockPos())){
+						BlockState state = world.getBlockState(scheduledLeafBreak.getBlockPos());
+						Block block = state.getBlock();
+						if(BlockTags.LEAVES.contains(block)){
+							block.randomTick(state, world, scheduledLeafBreak.getBlockPos(), world.getRandom());
+						}
+						else{
+							leavesBreak.remove();
+						}
+					}
+					else{
+						leavesBreak.remove();
+					}
+				}
+				else{
+					scheduledLeafBreak.tick();
 				}
 			}
 		}