+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.15.2-2.0.4] - 2020-02-29
+- Delay breaking leaves by a few ticks making (fixes #6)
+
 ## [1.15.2-2.0.3] - 2020-01-28
 - Added an option to force destroy leaves withing a certain radius. This will be applied from one of the top most log blocks and will destroy all leaves within it (including leaves that shouldn't despawn because another tree is too close, or because they've been placed by a player). (#5)